# swagger.yaml
openapi: 3.0.3
info:
  title: BanteayDigital API
  version: 3.0.0
  description: |
    Users privately scan suspicious content, explicitly request moderation, and
    administrators separately publish sanitized community safety alerts. The current
    scan engine is deterministic; Qdrant retrieval and grounded AI extend this scan contract.
servers:
  - url: http://localhost:3000
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
    cookieAuth: { type: apiKey, in: cookie, name: token }
  parameters:
...
    Page: { name: page, in: query, schema: { type: integer, minimum: 1, default: 1 } }
    Limit: { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
...
  schemas:
    ScanInput:
      type: object
      required: [inputType, value]
      properties:
        inputType: { type: string, enum: [TEXT, URL] }
        value: { type: string, minLength: 1, maxLength: 10000 }
    ImageScanInput:
      type: object
      required: [inputType, image]
      properties:
        inputType: { type: string, enum: [IMAGE] }
        image: { type: string, format: binary, description: PNG, JPG, JPEG, or WEBP; maximum 10 MB. }
    Scan:
      type: object
      required: [id, inputType, rawInput, normalizedInput, findings, assessment, score, createdAt]
      properties:
        id: { type: string }
        inputType: { type: string, enum: [TEXT, URL, IMAGE] }
        imageStoragePath:
          type: string
          nullable: true
          description: Private Supabase Storage object path; present only for IMAGE scans.
        imageMimeType: { type: string, nullable: true, example: image/png }
        imageSize: { type: integer, nullable: true, minimum: 0 }
        rawInput: { type: string }
        normalizedInput: { type: string }
        findings: { type: array, items: { type: object } }
        assessment:
          type: string
          description: No warning signs is not a guarantee that content is safe.
          enum: [NO_STRONG_WARNING_SIGNS, INSUFFICIENT_EVIDENCE, CAUTION, SUSPICIOUS, STRONG_SCAM_INDICATORS, UNABLE_TO_ASSESS]
        score: { type: integer, minimum: 0 }
        analysisSummary: { type: string, nullable: true }
        matchedScamCases: { type: array, items: { type: object } }
        recommendations: { type: array, items: { type: string } }
        createdAt: { type: string, format: date-time }
...
    SafetyKnowledgeInput:
      type: object
      required: [title, slug, category, shortDescription, content, warningSigns, preventionTips, indicators]
      properties:
        title: { type: string, minLength: 3, maxLength: 255, example: Recognize phishing links }
        slug: { type: string, pattern: '^[a-z0-9]+(?:-[a-z0-9]+)*$', example: recognize-phishing-links }
        category: { type: string, maxLength: 100, example: Phishing }
        shortDescription: { type: string, minLength: 10, maxLength: 500 }
        content: { type: string, minLength: 10, maxLength: 50000 }
        warningSigns: { type: array, items: { type: string }, minItems: 1 }
        preventionTips: { type: array, items: { type: string }, minItems: 1 }
        indicators: { type: array, items: { type: string }, minItems: 1, description: Retained for future scan recommendations. }
        relatedTopicIds: { type: array, items: { type: string }, maxItems: 20 }
        imageUrl: { type: string, format: uri, nullable: true }
        icon: { type: string, nullable: true, example: shield }
        isPublished: { type: boolean, default: false }
...
  responses:
    BadRequest: { description: Invalid request data. }
    Unauthorized: { description: Authentication is required or invalid. }
    Forbidden: { description: Administrator permission is required. }
...
    Conflict: { description: Action conflicts with resource state. }
paths:
...
  /api/v1/auth/me:
    get:
      tags: [Authentication]
      security: [{ bearerAuth: [] }]
      summary: Get the authenticated user
      responses: { '200': { description: User found. }, '401': { $ref: '#/components/responses/Unauthorized' } }
  /api/v1/auth/logout:
    post:
      tags: [Authentication]
      security: [{ bearerAuth: [] }]
      summary: End the current session
      responses: { '200': { description: Logged out. }, '401': { $ref: '#/components/responses/Unauthorized' } }
  /api/v1/scans:
    post:
      tags: [Scans]
      security: [{ bearerAuth: [] }]
      summary: Privately scan text, a URL, or an image
      description: |
        Choose one input type. Send TEXT or URL as JSON. Send IMAGE as
        multipart/form-data with an `image` file; its OCR text follows the same
        deterministic, retrieval, and AI-analysis flow as TEXT.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ScanInput' }
          multipart/form-data:
            schema: { $ref: '#/components/schemas/ImageScanInput' }
      responses:
        '201':
          description: Scan completed.
          content:
            application/json:
              schema:
                type: object
                required: [scan]
                properties:
                  scan: { $ref: '#/components/schemas/Scan' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '413': { description: Image exceeds the 10 MB upload limit. }
        '415': { description: IMAGE input must be PNG, JPG, JPEG, or WEBP. }
        '422': { description: Image OCR failed or no readable text was found. }
        '503': { description: AI OCR, retrieval, or analysis service is unavailable. }
...
  /api/v1/reports:
    get:
      tags: [Reports]
      security: [{ bearerAuth: [] }]
      summary: List the current user's reports
      parameters: [{ $ref: '#/components/parameters/Page' }, { $ref: '#/components/parameters/Limit' }]
      responses: { '200': { description: Reports found. }, '401': { $ref: '#/components/responses/Unauthorized' } }
...
  /api/v1/admin/reports:
    get:
      tags: [Admin Reports]
      security: [{ bearerAuth: [] }]
      summary: List reports for moderation
      parameters: [{ $ref: '#/components/parameters/Page' }, { $ref: '#/components/parameters/Limit' }]
      responses: { '200': { description: Reports found. }, '401': { $ref: '#/components/responses/Unauthorized' }, '403': { $ref: '#/components/responses/Forbidden' } }
...
  /api/v1/admin/users:
    get:
      tags: [Admin Console]
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      summary: List standard users and their report activity
      parameters:
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/Limit' }
        - { name: search, in: query, schema: { type: string, maxLength: 200 } }
        - { name: status, in: query, schema: { type: string, enum: [ACTIVE, SUSPENDED, BANNED] } }
      responses: { '200': { description: Users found. }, '401': { $ref: '#/components/responses/Unauthorized' }, '403': { $ref: '#/components/responses/Forbidden' } }
...
  /api/v1/safety-knowledge:
    get:
      tags: [Community Safety]
      summary: List published Community Safety knowledge
      description: Only published knowledge and published related topics are returned.
      parameters:
        - { name: category, in: query, schema: { type: string } }
        - { name: q, in: query, schema: { type: string } }
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/Limit' }
      responses: { '200': { description: Published knowledge found. } }
...
  /api/v1/admin/safety-knowledge:
    get:
      tags: [Admin Community Safety]
      security: [{ bearerAuth: [] }]
      summary: List all Community Safety knowledge, including drafts
      parameters:
        - { name: category, in: query, schema: { type: string } }
        - { name: q, in: query, schema: { type: string } }
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/Limit' }
      responses: { '200': { description: Knowledge found. }, '401': { $ref: '#/components/responses/Unauthorized' }, '403': { $ref: '#/components/responses/Forbidden' } }
    post:
      tags: [Admin Community Safety]
      security: [{ bearerAuth: [] }]
      summary: Create Community Safety knowledge
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/SafetyKnowledgeInput' } } } }
      responses: { '201': { description: Knowledge created. }, '400': { $ref: '#/components/responses/BadRequest' }, '409': { $ref: '#/components/responses/Conflict' } }
...